refactor: modify package structure
