+++ bandlog-user/src/main/java/net/effize/bandlog/user/model/Nickname.java
@@ -1,4 +1,4 @@
-package net.effize.bandlog.domain.user;
+package net.effize.bandlog.user.model;
 
 import jakarta.persistence.Embeddable;
